Put file URL updating code into a new component class
